Separate client and server directories
