+++ lib/client/doge.js
