+++ lib/client/event-stream.js
