+++ lib/client/index.js
