+++ lib/client/rendering.js
