+++ lib/client/stage.js
