+++ lib/server/palette.js
