+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "build": "browserify lib/client.js -t babelify -o client.js",
+    "build": "browserify lib/client/index.js -t babelify -o client.js",
     "dev": "nodemon -w lib -e js -x 'npm run build'",
     "start": "node_modules/.bin/babel-node server.js"
   },
